Refactored renderer texture loading.
Moved the texel decoding functions into the texture loading functions since the compiler wasn't appearing to inline the texel functions and was causing it to take way longer than it should. Also removed hardcoded dimensions for VoxelTexture and ChasmTexture (although VoxelTexture will most likely be forced power-of-2 for mipmaps once that is a thing). This new loading pattern also makes it easier for textures to do whatever custom laying-out they need in the future.