Redesigned EntityDefinition types.
Now have eight types of entity definitions for all the different kinds of entities in the world. This should make it easier to fully capture all of the original game's entities and classify them cleanly. Next, need to make a function that converts flatIndex to EntityDefinition::Type so shopkeepers, etc. can be static NPCs instead of doodads.