Start using EntityDefinitionLibrary.
Unfortunately human enemies aren't supported in the library because their corpse texture depends on the current level's .INF file, despite the texture probably being the same for every .INF file. Only creatures and citizens are supported for now, which means that they can be spawned in any level without needing to know anything about that level, in theory.