Added MapGeneration functions.
Copied over a lot of the interior Arena voxel loading from LevelData and it revealed that the FLOR and MAP1 loaders were deferring FLAT loading until the LevelData setActive(), which is interesting. Will probably want some kind of InteriorDefinition to hide the isPalace and rulerIsMale parameters in the interior level loading.