Redesigned TextureManager for TextureBuilders.
The Renderer drawing API is currently broken with several not-implemented macro placeholders and will need to be changed to internally allocate Texture instances for TextureBuilderID + PaletteID pairs as an interim solution to the renderer returning allocated texture handles. Much less of the codebase makes assumptions about textures being 8-bit now, which is a big improvement.