Revised EntityAnimationInstance data.
Keyframes don't have any data for now, and there is now an optional CitizenParams member that should only be allocated if the animation instance is for a citizen. Once the entity texture handles are implemented in the renderer, then it will make sense to have keyframes store those handles, and it will be easier to implement enemy NPC weapon frames.