SoftwareRenderer: Fix data race on flats.threadsDone
When the last render thread notify that flats are drawn, the main thread
have time to loop on and init flats again before others render threads
have time to do their own checks and continue, so they stay waiting
because flats.threadsDone was init to 0 again.

Remove the wait is safe because the first render threads to finish will
wait on the threadData.go already to false.