Manually tick level chunks once during initialization.
When fast travelling to a world map location, the new level instance hadn't been updated that frame yet, so the renderer was being passed an empty level on the first frame. Not sure if this patch is putting off a larger fix in the vein of deferring entity deletion until the beginning of the next frame.