Added exclusive fullscreen support.
It seems to work in general. Only two oddities so far: first is that if any SDL_Texture tries to update when exclusive fullscreen is on but the game is not focused, the lock texture call fails. Second is that starting the engine in exclusive fullscreen and then switching to windowed will make the window appear like borderless fullscreen because the renderer wasn't initialized with the dimensions from the options (maybe an easy fix).