Removed existing software renderer implementation.
Tried salvaging a few functions and put them in LegacyRendererUtils. I thought about keeping more of the old rendering code but it really was far too tightly-coupled with VoxelDefinitions and texture assets and things. Will be nice once the rendering is full 3D and is working with screen tiles and pixel shader threads.