Added lateTick() to attempt to fix some map transition issues.
Need to have some kind of level transition queue processed at a single point in the game loop. Currently we can do a LevelInstance::trySetActive() in handleInput() (clicking a voxel), tick() (walking into a voxel), and now lateTick() (finishing fast travel animation). There is now an issue with popping level instances from the stack, as it seems the dirty voxels in chunks are missed, resulting in no voxel geometry for the destination level.