More WIP decoupling weather from map generation.
Setting this aside since it's still a mountain of work to change the MapGeneration functions to not all pair with INFFile or LevelInfoDefinition. LevelDefinition should decide unique def ID allocation and sharing, and should probably take the weather-independent defs that LevelInfoDefinition has. LevelInfoDefinition should simply provide defs in a way similar to how VoxelTextureDefinition provides textures.