Added floor replacement defs and IDs to level.
It's better now having them generated in MapGeneration and inserted into LevelDefinition/LevelInfoDefinition. Hit a snag when testing Mines of Khuras because the VoxelChunk decorator populate code skips defs if they're outside the chunk writing range, and that was causing the floor replacement IDs to be wrong, but it's fixed.