WIP replacing MapInstance with SceneManager.
Still needs work for setting the active map and populating chunk managers correctly, but this should make the engine more understandable because the scene manager represents the active scene and there is no longer a list of potentially-active scenes like there was with MapInstance. GameState stores the MapDefinition and active level index and will be treated as the source of what the scene gets populated with.