Added per-pixel and per-mesh lighting types.
Also fixed how various objects calculate light and fixed lighting for objects using the light level transparency shader. Glad that "per-mesh" lighting is now a thing since it saves a lot of shading cost. Ironic that the Opaque pixel shader is now the most expensive overall because it is the most common and needs per-pixel lighting all the time.