Support spawning arbitrary entities.
For a long time, apart from citizens, the engine has only allowed entities to be from a LevelDefinition. Thinking of going in the direction where entities don't have to be registered to a chunk. Chunks will be more of an extension, and the entity manager will just iterate through all of stuff instead of doing for each chunk.