Added multipass support for stars.
Took a few attempts before figuring out how to get the sky mesh into the target color buffer for stars to draw onto. Had to lean on RenderDoc some. Unfortunately the ping-ponging pattern seems to require a copy of the sampled framebuffer to the target framebuffer because the sky mesh isn't writing out the same color to two color attachments (kind of a lesser of two evils thing?).