Fixed stars in front of planets in Vulkan.
The actual problem was that the software renderer was drawing the planet to the framebuffer first (strangely) then the star was trying to draw after but failed the previous brightness test, while in Vulkan both of those shaders read from the same sky mesh framebuffer. So now this problem's moved from Vulkan to the software renderer and the software renderer probably needs to ping-pong framebuffers.